perf: add runtime SIMD dispatch
Detect CPU SIMD capabilities at runtime and route vector space, fastscan, warmup, rotator, and excode packing paths through cached dispatch targets.

Split AVX2 and AVX512 implementations into ISA-specific translation units with scoped compile flags, including AVX512 popcnt handling for HNSW search.

Keep scalar public headers portable while adding coverage for CPU feature detection and dispatched distance behavior.
