+++ include/rabitqlib/index/estimator.hpp
@@ -128,6 +128,35 @@ inline float split_distance_boosting(
     return ex_dist;
 }
 
+template <class Kernel, class Query>
+inline void split_single_fulldist_direct(
+    const char* bin_data,
+    const char* ex_data,
+    float (*ip_func_)(const float*, const uint8_t*, size_t),
+    const Query& q_obj,
+    size_t padded_dim,
+    size_t ex_bits,
+    float& est_dist,
+    float& low_dist,
+    float& ip_x0_qr,
+    float g_add,
+    float g_error
+) {
+    ConstBinDataMap<float> cur_bin(bin_data, padded_dim);
+    ConstExDataMap<float> cur_ex(ex_data, padded_dim, ex_bits);
+
+    // [TODO: optimize this function]
+    ip_x0_qr = Kernel::mask_ip_x0_q(q_obj.rotated_query(), cur_bin.bin_code(), padded_dim);
+
+    est_dist =
+        cur_ex.f_add_ex() + g_add +
+        (cur_ex.f_rescale_ex() *
+         (static_cast<float>(1 << ex_bits) * ip_x0_qr +
+          ip_func_(q_obj.rotated_query(), cur_ex.ex_code(), padded_dim) + q_obj.kbxsumq()));
+
+    low_dist = est_dist - (cur_bin.f_error() * g_error / static_cast<float>(1 << ex_bits));
+}
+
 /**
  * @brief Batch distance estimation for qg. Here, we do not need intermediate results and
  * lower bound
@@ -200,6 +229,34 @@ inline void split_single_estdist(
     ConstBinDataMap<float> cur_bin(bin_data, padded_dim);
 
     ip_x0_qr = warmup_ip_x0_q_512(
+        cur_bin.bin_code(),
+        q_obj.query_bin(),
+        q_obj.delta(),
+        q_obj.vl(),
+        padded_dim,
+        q_obj.num_bits()
+    );
+
+    est_dist =
+        cur_bin.f_add() + g_add + (cur_bin.f_rescale() * (ip_x0_qr + q_obj.k1xsumq()));
+
+    low_dist = est_dist - (cur_bin.f_error() * g_error);
+};
+
+template <class Kernel>
+inline void split_single_estdist_direct(
+    const char* bin_data,
+    const SplitSingleQuery<float>& q_obj,
+    size_t padded_dim,
+    float& ip_x0_qr,
+    float& est_dist,
+    float& low_dist,
+    float g_add = 0,
+    float g_error = 0
+) {
+    ConstBinDataMap<float> cur_bin(bin_data, padded_dim);
+
+    ip_x0_qr = Kernel::warmup_ip_x0_q_512(
         cur_bin.bin_code(),
         q_obj.query_bin(),
         q_obj.delta(),
