+++ include/rabitqlib/index/ivf/ivf.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <immintrin.h>
 #include <omp.h>
 
 #include <algorithm>
