+++ include/rabitqlib/utils/memory.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
+#if defined(__x86_64__) || defined(__i386__)
 #include <immintrin.h>
+#endif
 #include <sys/mman.h>
 
 #include <cstdlib>
