+++ sample/cpp/CMakeLists.txt
@@ -8,3 +8,15 @@ add_executable(ivf_rabitq_querying ivf_rabitq_querying.cpp)
 
 add_executable(hnsw_rabitq_indexing hnsw_rabitq_indexing.cpp)
 add_executable(hnsw_rabitq_querying hnsw_rabitq_querying.cpp)
+
+foreach(RABITQ_SAMPLE_TARGET
+    symqg_indexing
+    symqg_querying
+    ivf_rabitq_indexing
+    ivf_rabitq_querying
+    hnsw_rabitq_indexing
+    hnsw_rabitq_querying
+)
+    target_link_libraries(${RABITQ_SAMPLE_TARGET} PRIVATE rabitq_headers)
+    target_compile_options(${RABITQ_SAMPLE_TARGET} PRIVATE -march=native)
+endforeach()
