+++ tests/unit/rabitqlib/utils/space_test.cpp
@@ -9,42 +9,80 @@
 using namespace rabitqlib;
 using namespace rabitq_test;
 
-TEST(Select_IP_Func, returns_correct_function_pointer) {
+TEST(Select_IP_Func, returns_stable_function_pointer) {
     auto ip_func = select_excode_ipfunc(0);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip16_fxu1_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(0));
 
     ip_func = select_excode_ipfunc(1);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip16_fxu1_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(1));
 
     ip_func = select_excode_ipfunc(2);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip64_fxu2_avx);
-    
+    ASSERT_EQ(ip_func, select_excode_ipfunc(2));
+
     ip_func = select_excode_ipfunc(3);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip64_fxu3_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(3));
 
     ip_func = select_excode_ipfunc(4);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip16_fxu4_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(4));
 
     ip_func = select_excode_ipfunc(5);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip64_fxu5_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(5));
 
     ip_func = select_excode_ipfunc(6);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip64_fxu6_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(6));
 
     ip_func = select_excode_ipfunc(7);
     ASSERT_NE(ip_func, nullptr);
-    ASSERT_EQ(ip_func, excode_ipimpl::ip64_fxu7_avx);
+    ASSERT_EQ(ip_func, select_excode_ipfunc(7));
 
     ip_func = select_excode_ipfunc(8);
     ASSERT_NE(ip_func, nullptr);
     ASSERT_EQ(ip_func, (excode_ipimpl::ip_fxi<float, uint8_t>));
+}
+
+TEST(ScalarQuantize, Uint8MatchesRoundedScalar) {
+    constexpr size_t dim = 37;
+    constexpr float lo = -3.0F;
+    constexpr float delta = 0.25F;
+    std::vector<float> input(dim);
+    std::vector<uint8_t> result(dim);
+    std::vector<uint8_t> expected(dim);
+
+    for (size_t i = 0; i < dim; ++i) {
+        float quantized = static_cast<float>((i * 7) % 251) + (static_cast<int>(i % 3) - 1) * 0.2F;
+        input[i] = lo + delta * quantized;
+        expected[i] = static_cast<uint8_t>(std::round((input[i] - lo) / delta));
+    }
+
+    scalar_quantize<uint8_t>(result.data(), input.data(), dim, lo, delta);
+
+    ASSERT_EQ(result, expected);
+}
+
+TEST(ScalarQuantize, Uint16MatchesRoundedScalar) {
+    constexpr size_t dim = 41;
+    constexpr float lo = 2.0F;
+    constexpr float delta = 0.125F;
+    std::vector<float> input(dim);
+    std::vector<uint16_t> result(dim);
+    std::vector<uint16_t> expected(dim);
+
+    for (size_t i = 0; i < dim; ++i) {
+        float quantized = static_cast<float>(1000 + i * 317) + (static_cast<int>(i % 5) - 2) * 0.1F;
+        input[i] = lo + delta * quantized;
+        expected[i] = static_cast<uint16_t>(std::round((input[i] - lo) / delta));
+    }
+
+    scalar_quantize<uint16_t>(result.data(), input.data(), dim, lo, delta);
+
+    ASSERT_EQ(result, expected);
 }
 
 TEST(ip16_fxu1_avx, ip_works) {
